requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.2
pandas==2.2.2
streamlit==1.35.0
python-dotenv==1.0.1
